+++ src/bin/osm-lump-ways-down/cli_args.rs
@@ -158,6 +158,16 @@ pub struct Args {
     #[arg(long, value_name = "OUTPUT.geojson[s]")]
     pub loops: Option<PathBuf>,
 
+    /// Whether to include a string list of all the consituant nids for each loop. For very long
+    /// loops, this can make a large GeoJSON file, which some tools (e.g. `ogr2ogr`) refuse to deal
+    /// with.
+    #[arg(long, requires="loops", conflicts_with="loops_no_incl_nids")]
+    pub loops_incl_nids: bool,
+
+    /// Don't include the `nids` property in the loops file
+    #[arg(long, requires="loops", conflicts_with="loops_incl_nids")]
+    pub loops_no_incl_nids: bool,
+
     /// Path to store CSV of statistics
     /// CSV file with 4 columns.
     /// • `timestamp`: unix epoch timestamp of data age (integer)